Move TermNav component into own file
