+++ styles/pages/_ClassPage.scss
@@ -230,7 +230,7 @@
     }
 
     .sectionCards .horizontalLine {
-      margin: 50px 0px;
+      margin: 60px 0px;
     }
 
     .sectionCard {
